+++ locale/ja_JP/LC_MESSAGES/messages.po
@@ -26,5 +26,5 @@ msgstr "雪"
 msgid "rain"
 msgstr "雨"
 
-
-
+msgid "Indoor"
+msgstr "室内"
